Fix: move helpers around
