+++ tmaze/src/game/activity.rs
@@ -1,0 +1,7 @@
+pub struct Activity {
+    // source, ie. mod or base game
+    source: String,
+    name: String,
+
+
+}
